Move all convertV1Profiles() interfaces into IConfigConvert
