--- packages/imperative/src/config/src/doc/IConfigConvertResult.ts
@@ -1,22 +1,0 @@
-/*
-* This program and the accompanying materials are made available under the terms of the
-* Eclipse Public License v2.0 which accompanies this distribution, and is available at
-* https://www.eclipse.org/legal/epl-v20.html
-*
-* SPDX-License-Identifier: EPL-2.0
-*
-* Copyright Contributors to the Zowe Project.
-*
-*/
-
-import { IConfig } from "./IConfig";
-
-export interface IConfigConvertResult {
-    config: IConfig;
-    profilesConverted: { [key: string]: string[] };
-    profilesFailed: {
-        name?: string;
-        type: string;
-        error: Error;
-    }[];
-}
